+++ condow_core/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.4] - unreleased
+
+### ADDED
+
+- `Reporter` can track failed parts
+- `Logger` logs failed parts
+
 ## [0.12.3] - 2022-02-07
 
 ### ADDED